Implementado serviço de fila para recuperação de senha com email utilizando RabbitMQ

+++ beiramar.api/src/main/java/com/beiramar/beiramar/api/repository/LogSenhaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LogSenhaRepository extends JpaRepository<LogSenhaEntity, Long> {
+public interface LogSenhaRepository extends JpaRepository<LogSenhaEntity, Integer> {
     Optional<LogSenhaEntity> findTopByTokenAndUsuarioEmailOrderByDataHoraLogSenhaDesc(String token, String email);
 }
